feat: add sqlite test:ci setup
